definiendo valores a double en los modelos

+++ app/Models/Car.php
@@ -20,6 +20,6 @@ class Car extends Model
     }
 
     protected $casts = [
-        'amount' => 'float'
+        'amount' => 'double'
     ];
 }
